refactor: 각각 엔티티에 BaseEntity를 추가한다. ---
refactor: 각각 엔티티에 BaseEntity를 추가한다.
