+++ src/main/java/com/fourweekdays/fourweekdays/Inbound/Inbound.java
@@ -1,9 +1,10 @@
 package com.fourweekdays.fourweekdays.Inbound;
 
+import com.fourweekdays.fourweekdays.common.BaseEntity;
 import jakarta.persistence.*;
 
 @Entity
-public class Inbound {
+public class Inbound extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
