+++ src/main/java/com/fourweekdays/fourweekdays/franchise/FranchiseStatus.java
@@ -1,0 +1,6 @@
+package com.fourweekdays.fourweekdays.franchise;
+
+public enum FranchiseStatus {
+    ACTIVE, INACTIVE
+}
+
