+++ src/main/java/com/fourweekdays/fourweekdays/notification/Notification.java
@@ -1,6 +1,7 @@
 package com.fourweekdays.fourweekdays.notification;
 
 import com.fourweekdays.fourweekdays.Inbound.Inbound;
+import com.fourweekdays.fourweekdays.common.BaseEntity;
 import com.fourweekdays.fourweekdays.inventory.Inventory;
 import com.fourweekdays.fourweekdays.member.Member;
 import com.fourweekdays.fourweekdays.outbound.Outbound;
@@ -8,7 +9,7 @@ import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-public class Notification {
+public class Notification extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
