+++ src/main/java/com/fourweekdays/fourweekdays/warehouse/Warehouse.java
@@ -1,10 +1,11 @@
 package com.fourweekdays.fourweekdays.warehouse;
 
 import com.fourweekdays.fourweekdays.common.Address;
+import com.fourweekdays.fourweekdays.common.BaseEntity;
 import jakarta.persistence.*;
 
 @Entity
-public class Warehouse {
+public class Warehouse extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
